tails-security-check-wrapper: Fix problematic call to inotifywait.
inotifywait immediately fails if the watched file doesn't exist which
results in .xession-errors getting spammed with errors occupying
precious space in the ramfs. Since $TORDATE_DIR is created by a
network-manager hook, this becomes a serious problem when Tails starts
without a network connection, but it always results in some amount of
lost RAM.

Since the security check script is run under the normal user we cannot
create it, so we make inotifywait fallback to sleep instead. Actually
we could just replace it with sleep altoghether...