Split VM networking into its own class, VMNet.
This split makes the setup of all the libvirt stuff on
feature/scenario start and end cleaner. Now we create one new VMNet
per @source-tagged feature, and let it live throughout that feature,
after which it is killed. This has some very nice reliablity and
performance advantages compared to creating a VMNet per scenario
(which would be equivalent to what we've been doing until now):

By design we want a new VM instance per scenario (possibly restored
from a snapshot), but that's not necessary for the VMNet instance; it
could just as well survive for the whole feature. One good reason for
the latter is that restoring a VM and have it use another VMNet
instance than it was saved from confuses libvirt in such a way that
re-establishing the network takes around 30 seconds longer compared to
restoring it to the same VMNet instance.