Fix privilege escalation after reboot
tails-additional-software accepted to load a packages list file
writable by non-root or living in a directory writable by non-root.
As a consequence, non-root user was able to trigger the installation of
any debian package in the next login with persistence activated.

This commit fixes this issue by refusing to load a conffile which
mode is not 644 or living in a directory which mode is not 755.