Migrant our Vagrant instance to the libvirt provider.
The most notable side effect of the migration is that we don't rely on
shared folders anymore since the libvirt provider uses rsync, which
only happens on VM start. Since the syncing isn't two-way, we cannot
rely on it for getting the build results, and since the syncing only
happens on start, we cannot `git checkout` another branch or commit
changes on the host after we've started the VM and expect the sources
inside the VM to be in the host's state.

To get the equivalent of the old behaviour we instead explicitly
upload the source for each build, and download the build results after
a successful build, using Vagrants download and upload primitives.