Actually run "Upgrade from ISO" from a USB drive running the old version.
That's what users do, and is buggy. If one already has an up-to-date Tails USB
stick, they can "Clone and upgrade", and won't bother using an ISO image.

Still, let's keep the "Clone from ISO while running the new version" test,
as it tests the Installer in the tested (new) version of Tails, which makes
little sense wrt. what users will do when the new version is out, but may catch
bugs that will affect upgrades to N+1.